lte: daemon: Fix bug that recvfrom() may not return immediately
Fixed a bug where recvfrom() may not return immediately
even if MSG_DONTWAIT was set.
When MSG_PEEK was specified with recvfrom(), lte_daemon
was notifying USRSOCK_EVENT_RECVFROM_AVAIL to usrsock.
Therefore, if recvfrom() is called with nothing in the
receive buffer of the protocol stack, recvfrom() will
not be returned until the broker disconnects.