bluetooth: nrf52: Set timeout value for connection
The timeout of scan parameters is used as the connection timeout
in BLE board.
Because the timeout value is set to 'infinitely'(timeout does not occur)
in scan start timing, the connection timeout setting is also 'infinitely'
without change.
So, modify this timeout value to finite value to avoid that application
waits connection infinitely.