sdk/modules/bluetooth/nrf52: Fix descriptor discovery bug
In case of last characteristic or UUID-specified discovery,
the range of descriptor handle can be invalid.
In such a case, all attribute data is notify and
characteristic group is delimited by UUID = 0x2803 information,
that is characteristic delaration.
So, Add 0x2803 delimitation codes to notify the correct information
to application.