examples/pospredict_pwbimu: Add an simple position predict sample
This is a sample code for simple self-position estimation by using
CXD5602PWDIMU sensor.
It estimates position by simply integrating the values of the
accelerometer and gyroscope over time. It is possible to clear
the velocity to zero with a key input, and by periodically creating
a zero velocity environment and clearing it to zero, you can feel
that drift can be suppressed to a certain extent.